requirements.txt
dill==0.3.5.1
matplotlib==3.4.3
numpy==1.21.4
pandas==1.5.0
Pillow==9.2.0
torch==1.12.0+cu116
torchvision==0.13.0+cu116
tqdm==4.62.3
